kaggle_requirements.txt
# Core PyTorch + Transformers
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
transformers==4.38.2
accelerate==1.5.2
datasets==3.4.1
sentence-transformers==4.0.1

# HuggingFace ecosystem
huggingface-hub==0.28.1
tokenizers==0.21.0
tiktoken==0.9.0

# LangChain & related tools
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-ollama==0.2.3
langchain-huggingface==0.1.2
langgraph==0.2.69
langsmith==0.3.4

# ChromaDB (vector store)
chromadb==0.6.3
chroma-hnswlib==0.7.6

# Parsing / NLP / utils
beautifulsoup4==4.13.3
lxml==5.3.1
html5lib==1.1
nltk==3.9.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
PyYAML==6.0.2
orjson==3.10.15
markdown-it-py==3.0.0
regex==2024.11.6

# Data processing
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
tqdm==4.67.1

# System / logging / utils
coloredlogs==15.0.1
humanfriendly==10.0
filetype==1.2.0
openai==1.66.5
requests==2.32.3
rich==13.9.4
importlib_metadata==8.5.0
wrapt==1.17.2
